feat(product.persistence) : 테이블 초기화 메서드, 엔티티 삭제 시 없으면 예외 호출 로직 작성

+++ src/main/java/gift/product/persistence/repository/ProductRepositoryImpl.java
@@ -32,6 +32,12 @@ public class ProductRepositoryImpl implements ProductRepository{
 
     @Override
     public Long deleteProductById(Long id) {
+        if(!productJpaRepository.existsById(id)) {
+            throw new NotFoundException(
+                ErrorCode.DB_NOT_FOUND,
+                "Product with id " + id + " not found"
+            );
+        }
         productJpaRepository.deleteById(id);
         return id;
     }
@@ -54,5 +60,10 @@ public class ProductRepositoryImpl implements ProductRepository{
                 Product::getId,
                 product -> product
             ));
+    }
+
+    @Override
+    public void deleteAll() {
+        productJpaRepository.deleteAll();
     }
 }
